# plan: prepare candidate index