--- a/plan/physical_plan_builder.go
+++ b/plan/physical_plan_builder.go
@@ -212,6 +212,7 @@ func (ds *DataSource) convert2PhysicalPlan(prop *requiredProp) (task, error) {
 	}
 	if !includeTableScan || len(ds.pushedDownConds) > 0 || len(prop.cols) > 0 {
 		for i, idx := range indices {
+			// TODO: We can also check if the prop matches the index columns.
 			if !ds.validIndices[i] && len(prop.cols) == 0 {
 				continue
 			}