--- a/plan/property_cols_prune.go
+++ b/plan/property_cols_prune.go
@@ -20,22 +20,22 @@ import (
 func (ds *DataSource) preparePossibleProperties() (result [][]*expression.Column) {
 	indices := ds.availableIndices.indices
 	includeTS := ds.availableIndices.includeTableScan
+	ds.validIndices = make([]bool, len(indices))
 	if includeTS {
 		col := ds.getPKIsHandleCol()
 		if col != nil {
 			result = append(result, []*expression.Column{col})
 		}
 		cols := expression.ExtractColumnsFromExpressions(make([]*expression.Column, 0, 10), ds.pushedDownConds, nil)
-		ds.validIndices = make([]bool, len(indices))
-		for _, col := range cols {
-			for i, idx := range indices {
-				if !ds.validIndices[i] && col.ColName.L == idx.Columns[0].Name.L {
+		for i, idx := range indices {
+			for _, col := range cols {
+				if col.ColName.L == idx.Columns[0].Name.L {
 					ds.validIndices[i] = true
+					break
 				}
 			}
 		}
 	} else {
-		ds.validIndices = make([]bool, len(indices))
 		for i := range ds.validIndices {
 			ds.validIndices[i] = true
 		}
@@ -55,7 +55,9 @@ func (p *LogicalSelection) preparePossibleProperties() (result [][]*expression.C
 
 func (p *baseLogicalPlan) preparePossibleProperties() [][]*expression.Column {
 	if len(p.children) > 0 {
-		p.children[0].preparePossibleProperties()
+		for _, ch := range p.children {
+			ch.preparePossibleProperties()
+		}
 	}
 	return nil
 }