@coocood 2018-02-07
@hanfei1991 
If the index is a covering index and there is no push-down condition, will it be excluded?
And if the index order is needed, but there is no push-down condition, will it be excluded?