@hanfei1991 2018-02-07
@coocood According to https://github.com/pingcap/tidb/pull/5809/files#diff-fdcfa025d622ba63a631918a93b05d32R216 , we omit an index only when order columns is empty