@hanfei1991 2018-02-07
/run-common-test -tidb-test=pr/459